Implemented feature to fill budget progressbar according to total spent.

+++ app/src/main/java/com/example/android/travelwallet/model/TravelValues.java
@@ -1,5 +1,7 @@
 package com.example.android.travelwallet.model;
 
+import java.math.BigDecimal;
+
 public class TravelValues {
-    public double total;
+    public BigDecimal total;
 }
